requirements.txt
Django==4.0.4
django-tables2==2.4.1
django-extensions==3.1.3
django-filter==21.1
django-crispy-forms==1.13.0
django-spurl==0.6.8
biopython==1.79
psutil==5.8.0
mysqlclient==2.0.3
pandas==1.3.3
tablib==3.2.1
django-autocomplete-light==3.9.4
rpy2==3.5.2
django-secrets==1.2.4
